Moving test to its own file
