+++ crates/base/tests/null_body_status_null_body_tests.rs
@@ -71,21 +71,3 @@ async fn test_null_body_with_204_status_post() {
 
     assert_eq!(body_bytes.len(), 0);
 }
-
-#[tokio::test]
-async fn test_custom_readable_stream_response() {
-    select! {
-        _ = start_server(
-            "0.0.0.0",
-            8999,
-            String::from("./test_cases/main"),
-            None,
-            false
-        ) => {
-            panic!("This one should not end first");
-        }
-        resp = reqwest::get("http://localhost:8999/readable-stream-resp") => {
-            assert_eq!(resp.unwrap().text().await.unwrap(), "Hello world from streams");
-        }
-    }
-}
